Fixed Broken Checksum Routine & Restored Checksum Error Branch
Aurora made a silly mistake during the original implementation of the driver & forgot that the "ErrorTrap" label is now near the end of the ROM in the new error handler code. As such, the checksum routine would always thrown an error screen because it was only calculating the checksum for a very small section of ROM, which did not match the checksum of the full ROM in the header. As such, the error branch was commented out to make things work.

Changing the refrence from "ErrorTrap" to "EntryPoint" restores the proper functionality of the checksum routine, so it works properly now. Joy!